requirements.txt
pyodbc
gitpython
requests
cryptography
coverage
pytest
